Access controller and head acceleration and velocity from gdscript
This gdscript API extension allows to access the predicted acceleration and velocity
of the tracked controllers and the head.
It caches the controller pose update that is done per
frame and allows to access it via gdscript API functions. This data is a lot more
accurate then computing these values over multiple frames inside godot.