add hand pose tracking api and hand model
this adds a new .gdns api ovr_hand_tracking that allows to query the
hand pose from the VrApi when hand tracking is enabled. It also includes
a left and right hand rigged model and example code on how to map the
tracked hand pose from the API to the hand mesh. The example code is
part of the example_scenes.

mapping pinch gestures to ARVRController buttons/axis

We take the pinch gesture values from the SDK and map them to the
ARVRController button presses and axis. This allows some easy
interaction integration just using the ARVRController with hand
tracking without the need to call specific VrApi functions.

return hand tracking confidence values for pose

The VrApi gives a status about if the hands are currently tracked or not
and how confident the pose estimation is (only the values LOW and HIGH)
at the moment. This change returns for get_hand_pose a float value that
is 0.0 if the hand is not tracked; 0.5 for low and 1.0 for high
confidence. This allows users to fade/hide the hand models when the
tracking quality is not good.

allow runtime switch of controller type

this change checks if the type of the controller changed at runtime and
disconnects the old one and reconnects the new type. This allows runtime
switching between hands and oculus touch controllers without restarting
the application and will allow users to react to it in code by checking
the controller name.

pinch to button mapping

The new mapping is X, Y,
GripTrigger, IndexTrigger. This follows the position on the controllers
to the order of the fingers (Index, Middle, Ring, Pinky)

added demo/export_presets.cfg to .gitignore
updated README.md to require SDK 12