more API examples and doc improvements
Included ovrVrApiTypes loading example, a line on how to access the tracking bounding box and
examples on how to get the predicted acceleration and velocity values.
Clarified why there is an enum for the button mappings even though some of the buttons are
defined in godot core.
Expanded a bit the existing comments.